aura-agent/requirements.txt
# Aura IoT Troubleshooter Agent Requirements
# Python dependencies for the project

# Core LangChain and LangGraph - Using compatible versions
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
langsmith>=0.1.0

# OpenAI Integration
openai>=1.0.0

# AWS Integration
boto3>=1.34.0
botocore>=1.34.0

# Vector Database and Embeddings
chromadb>=0.4.22
sentence-transformers>=2.2.0

# PostgreSQL with pgvector
psycopg2-binary>=2.9.0
pgvector>=0.2.0

# Streamlit for web interface
streamlit>=1.30.0

# Document Processing
pypdf>=3.17.0
python-docx>=1.1.0
markdown>=3.5.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0

# Environment and Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# CLI and Utilities
click>=8.1.0
rich>=13.7.0
typer>=0.9.0

# Logging and Monitoring
structlog>=23.2.0

# HTTP and API
requests>=2.31.0
httpx>=0.26.0

# Development Dependencies
pytest>=7.4.0
black>=23.12.0
flake8>=7.0.0
mypy>=1.8.0

# Azure Vision (for future image analysis features)
azure-ai-vision-imageanalysis
Pillow # For image handling